test: make WeakRef finalization test more robust
The fact that the call stack was deep when the object was
created does not ensure that it will be finalized. Do
a deep call stack again to erase the old value.

Moreover, _id2ref is not a proper check to know if an object
has been finalized, since it is possible it has been reused.
Check for both cases.