fix detection of finalized WeakRef in do_object_finalize
in this function, we must detect WeakRef that are currently
being finalized (i.e. sequences where the object is finalized
and then the weakref pointing to it). The method used with 1.8
is not proper for 1.9. We should just check if the object has
FL_MARK set, which is true for both 1.8 and 1.9