fix a crash in Utilrb::WeakRef with recent versions of 1.8.7
The main issue is that Ruby does GC sweep pass and only then
calls the finalizers.

We were using FL_TEST to check if the weakref is being finalized
(in which case we should not access it anymore because it is
already freed), or if is still alive.

The issue came from the fact that we were using two finalizers:
one for the weakref and one for the object. The first change in
this commit is to remove the weakref's finalizer, and do the
cleanup in the free-function. That way, we actually know in the
object finalizer that a referenced weakref is alive.

BUT recent patchlevels in 1.8.7 (174 is broken) have also moved
the RDATA's freeing stage to after the sweep, I guess so that
the free functions can allocate objects (which is forbidden
during garbage collection). But they do so while resetting the
'flags' field of the Ruby object definition, breaking the object's
type.

Anyway, it just means that the object's finalizer MUST access
the data structure using RDATA and not Data_Wrap_Struct