module: add handling for a ModuleExtension special submodule in Module#include
The problem, in modules, is to separate between what should be available at the
instance level, at the class level and what we want to extend the other's
modules' with.

Until now, we had the plain module and mod::ClassExtension. The first one
extends the object instances and the second one extends the object classes. What
was missing is a way to gather methods that would extend the other modules.

This commit adds support for a ModuleExtension module. The following example:

  module Mod1
    module ModuleExtension
    end
    module ClassExtension
    end
  end
  module Mod2
    include Mod1
  end
  class K
    include Mod2
  end

Is equivalent to:

  Mod2.include Mod1
  Mod2.extend Mod1::ModuleExtension

  K.include Mod2
  # K.include Mod1 is handled by the previous line
  K.extend Mod1::ClassExtension
  # This is actually K.extend Mod2::ClassExtension, in which
  # Mod1::ClassExtension has been included by Mod2.include Mod1