add Kernel#with_modules and Object#scoped_eval
The latter (instance|module|class)_eval its given
block with the constant lookup rules of 1.8,
regardless of the interpreter version.

The reason is that this lookup rule is more DSL-friendly

Kerne#with_modules is injecting a module from which
constants should be looked up in the resolution chain.

It comes from

http://coderrr.wordpress.com/2009/06/02/fixing-constant-lookup-in-dsls-in-ruby-1-9/