module: add const_defined_here?
const_defined? behaves differently in 1.8 and 1.9. In 1.8, it will
return false if the constant is NOT defined in the receiver, but
is accessible from the receiver's context. In 1.9 it will return
true by default in this case, one needs to give "false" as a second
argument to turn on the 1.8 behaviour again.

const_defined_here? does that