define Marshal.load_with_missing_constants
The problem with Ruby's Marshal implementation is that it requires
all classes that have been dumped to be present at loading time.
This is dangerous as it makes loading old log files difficult at best.

This new method dynamically defines missing classes so that the
marshalled file can be loaded. Obviously, this should be used
only for data logs, as a code that would use this data is most likely
going to need the real classes.