rdoc: don't specify a full path for included and excluded files for RDoc
Backport from master/next. The reason is that this bug makes the autoproj
documentation rebuild the whole stack, which is a grave functionality bug.

The RDoc rake task has a bug, in which providing absolute paths
for the include and exclude sets causes the html file to be (wrongly)
generated in the source directory instead of in the output directory.