implement a workaround for Qt::Variant.from_value in qtruby
The issue with Qt::Variant.from_value(Object.new) is that it breaks. Often.

This method is pure-ruby, does not rely on anything fancy, but requires a bit
of thinking, as one must give a proper lifetime object to keep the ruby object
alive.

Note that if the goal is to get ruby objects across signal/slots, there is no
need for a lifetime object, as the QVariant will be this lifetime object.