event_loop: propagate the backtrace of the original exception when modifying it in the callback
Callbacks can return an exception object to override the error raised
by the work block. However, setting the backtrace of that exception
object is awkward, and will probably be forgotten most of the time.

Make sure that, if no backtrace has been provided, we propagate the
backtrace of the original exception into the new one.