thread_pool: rewrite the test suite
This is a complete rewrite of the test suite where all sleeps have
been removed, and logic is now mostly deterministic. In addition,
some tests that were only half-testing functionality because of
non-determinism (e.g. the sync key synchronization) are now testing
the outcome properly.

This is only the test suite, the rest of the commits will be a broken-down
set of changes which make the new test suite pass properly