thread_pool: refactor the main thread loop
Complex logic in blocks is really really hard to follow because
of the inability to use return. In this case, I barely was able to
follow the thread cancelling logic.

This commit does not change anything but moving code around. I've
done it here first to be able to comment on all changes I've made
to the actual logic.