event_loop: add the queue option to async and periodic calls
It controls whether the work should be queued right away, or if it
should either wait for a period (in the periodic every and
async_every calls) or be explicitly queued (in defer).

In addition, in async_every, it creates the thread pool task right
away so that it is available for the caller to manipulate (and e.g.
execute explicitely with #execute). This removes a delay in queueing
the work as we don't wait for #step to be called to feed the work
to the thread pool.