event_loop: make Forward raise right away if the accessor is not available even in the deferred case
Until now, a deferred method call would not check for the availability
of the accessor when the call is queued, only when it is made. However,
since there is virtually no way of reliably setting the accessor in
between, also do the check early. The late check is kept for safety
reasons (as it *is* possible to unreliably set it ;-))