thread_pool: fix synchronization when checking if a task is being used in #<<
The current code was testing against started? without taking any
mutex. This had two flaws: it was possible to queue the same
task twice -- probably with bad consequences -- and it was
possible for a task to be started between the check and the
queueing.

This replaces the reset method by an acquire method. Acquire
both sets the queued_at attribute *and* checks whether the task
is already in use or not within the same mutex, ensuring proper
synchronization.