event_loop: refactor step into smaller processing methods
This clarifies the step processing by splitting the code in smaller
steps.

It also removes the @number_of_events_to_process mechanism as IMO
it lost its purpose: either @number_of_events_to_process was greater
than @queue.size and it would block, or it is smaller and therefore
useless. Since the blocking behaviour is definitely unwanted, its
only purpose was to count processed events so that we don't process
newly queued events. This is handled by a boolean parameter to
process_events