thread_pool: make #sync_task execute the task once after an already running one
There is no guarantee that the currently running task will provide
the same work than a new one (i.e. the point in time at which the
task was queued might be important). So, run it another time to
ensure that one got executed *after* the call to #sync_task