refactor #check_arity
This expands the unit tests to cover all cases of all callable
objects (methods, procs and lambdas), and fixes inconsistencies.

It also adds a 'strict' keyword argument that allows to check against
method rules, that is that passing less or more arguments than the
object expects is forbidden (unlike with procs).